docs(*): relocate stories, fix cypress
