+++ packages/wix-storybook-utils/package.json
@@ -19,15 +19,13 @@
   "homepage": "https://github.com/wix/wix-ui#readme",
   "main": "./dist/src/index.js",
   "scripts": {
+    "prebuild": "yoshi lint",
     "build": "yoshi build && import-path --path src",
     "pretest": "build-storybook",
     "test": "npm run test:unit && npm run test:e2e",
-    "posttest": "npm run lint",
-    "test:watch": "yoshi test --jest --watch",
     "test:unit": "yoshi test --jest",
-    "test:e2e": "concurrently 'servedir storybook-static 6006' 'wait-on http://localhost:6006 && cypress run' -k -s first",
-    "lint": "yoshi lint",
-    "start": "yoshi start --no-test & npm run storybook",
+    "test:e2e": "concurrently 'servedir storybook-static 6006' 'wait-on http://localhost:6006 && cypress run' -k --success last",
+    "start": "npm run storybook",
     "release": "yoshi release",
     "storybook": "start-storybook -p 6006"
   },
