+++ packages/wix-storybook-utils/stories/sections/story-with-mdx.story.js
@@ -1,13 +1,13 @@
 import * as React from 'react';
-import Component from './component';
-import { mdx } from '../src/Sections';
+import Component from '../component';
+import { mdx } from '../../src/Sections';
 
-import MDX from './examples/markdown.mdx';
+import MDX from '../examples/markdown.mdx';
 
 export default {
   category: 'Components',
   storyName: 'Component with section 2',
   component: Component,
-  componentPath: './component.js',
+  componentPath: '../component.js',
   sections: [mdx({ content: MDX })],
 };
