+++ packages/wix-storybook-utils/stories/sections/story-without-component.story.js
@@ -1,0 +1,13 @@
+import { description } from '../../src/Sections';
+
+const markdown = require('../examples/markdown-titles.md');
+
+export default {
+  category: 'Sections',
+  storyName: 'Composite component',
+  sections: [
+    description({
+      text: markdown,
+    }),
+  ],
+};
