--- packages/wix-storybook-utils/stories/story-without-component.story.js
@@ -1,13 +1,0 @@
-import { description } from '../src/Sections';
-
-const markdown = require('./examples/markdown-titles.md');
-
-export default {
-  category: 'Components',
-  storyName: 'Composite component',
-  sections: [
-    description({
-      text: markdown,
-    }),
-  ],
-};
